requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-wordpress-xmlrpc==2.3
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
tqdm==4.66.1
urllib3==2.0.7
python-magic==0.4.27